Split experiment JSON controller in two; repackage OT dump classes and sanitise test to get rid of a warning
